samples/python/agents/adk_facts/pyproject.toml
[project]
name = "adk-facts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.11",
    "starlette>=0.46.1",
    "uvicorn>=0.34.0",
    "click>=8.1.8",
    "google-adk>=1.0.0",
    "python-dotenv>=1.1.0",
    "asyncpg>=0.30.0",
    "google-cloud-alloydb-connector[asyncpg]>=1.9.0",
]
